updated map integration commands
