+++ map-integration/macrostrat/map_integration/commands/ingest.py
@@ -11,6 +11,7 @@ from rich.console import Console
 from rich.progress import Progress
 from sqlalchemy import text
 
+from .geodatabase import get_vector_info
 from ..database import db
 from ..errors import IngestError
 
@@ -152,8 +153,12 @@ def create_dataframe_for_layer(file: Path, layer: str) -> G.GeoDataFrame:
 
 
 def get_dataframes(files) -> Iterable[Tuple[str, G.GeoDataFrame]]:
+    single_file = len(files) == 1
     for file in files:
-        print(file)
+        console.print(file, style="bold cyan")
+
+        info = get_vector_info(file)
+
         with F.open(file) as f:
             print(f.driver)
             print(f.crs)
@@ -161,36 +166,50 @@ def get_dataframes(files) -> Iterable[Tuple[str, G.GeoDataFrame]]:
         layers = F.listlayers(file)
         n_layers = len(layers)
         if n_layers > 1:
-            console.print(f"{n_layers} layers found in {file}.")
+            console.print(f"{n_layers} layers.")
 
         for layer in layers:
             console.print(f"Layer: {layer}")
 
+            # Create the basic data frame
             df = G.read_file(file, layer=layer)
 
-            console.print(file, style="bold cyan")
+            _print_layer_info(df, console)
 
-            # If there is no geometry, skip
-            if "geometry" not in df.columns:
-                console.print("No geometry column found. Skipping.")
-                continue
-
-            # Print geometry type statistics
-            counts = df.geometry.type.value_counts()
-            for geom_type, count in counts.items():
-                console.print(f"- {count} {geom_type}s")
-
-            _col_list = ", ".join(df.columns)
-            # Print out column names
-            console.print(f"- [bold]Columns[/bold]: [dim]{_col_list}[/dim]")
-
-            name = file.stem
-            if n_layers > 1:
-                name += f"_{layer}"
-                if len(files) == 1:
-                    name = layer
+            name = get_layer_name(
+                file, layer, single_file=single_file, single_layer=n_layers == 1
+            )
 
             yield name, df
+
+
+def _print_layer_info(df, _console: Console):
+
+    # If there is no geometry, skip
+    if "geometry" not in df.columns:
+        _console.print("No geometry column found. Skipping.")
+        return
+
+    # Print geometry type statistics
+    counts = df.geometry.type.value_counts()
+    for geom_type, count in counts.items():
+        _console.print(f"- {count} {geom_type}s")
+
+    _col_list = ", ".join(df.columns)
+    # Print out column names
+    _console.print(f"- [bold]Columns[/bold]: [dim]{_col_list}[/dim]")
+
+
+def get_layer_name(
+    file: Path, layer: str, single_file=False, single_layer=False
+) -> str:
+    """Get the best layer name for a file and layer."""
+    name = file.stem
+    if single_layer:
+        return name
+    if single_file:
+        return layer
+    return f"_{layer}"
 
 
 def chunker(seq, size):
